JP-2460 - Migrations of communication history           whittle away at the compile errors

+++ comu/com.energyict.mdc.engine/src/main/java/com/energyict/mdc/engine/impl/commands/store/deviceactions/CreateMeterEventsFromStatusFlagsCommandImpl.java
@@ -78,7 +78,7 @@ public class CreateMeterEventsFromStatusFlagsCommandImpl extends SimpleComComman
                     }
                     collectedMeterEvents.addAll(MeterEvent.mapMeterEventsToMeterProtocolEvents(meterEvents));
                 }
-                DeviceIdentifierById deviceIdentifier = new DeviceIdentifierById(loadProfileCommand.getOfflineDevice().getId());
+                DeviceIdentifierById deviceIdentifier = new DeviceIdentifierById((int) loadProfileCommand.getOfflineDevice().getId());
                 DeviceLogBook deviceLogBook = new DeviceLogBook(new LogBookIdentifierByDeviceAndObisCode(deviceIdentifier, DeviceLogBook.GENERIC_LOGBOOK_TYPE_OBISCODE));
                 deviceLogBook.setMeterEvents(collectedMeterEvents);
                 this.loadProfileCommand.addCollectedDataItem(deviceLogBook);
